API: take SVG `viewBox` scaling into account
Inkscape returns the coordinates and bounding box
dimensions in pixels (SVG user units) measured
outside the SVG `viewBox` (global).

The paths and other elements within the `viewBox`
have dimensions in SVG user units within the `viewBox`
(local). So a conversion is needed from the pixels
inside the `viewBox` to the pixels outside the `viewBox`.

For `man inkscape` it follows that the dimensions of
`--query-all` are in (global) pixels:

> -X, --query-x
>    Query the X coordinate of the drawing or, if specified,
>    of the object with --query-id. The returned value is
>    in px (SVG user units).